image_processing/requirements.txt
google-generativeai
Pillow
speechrecognition
pyaudio
python-dotenv
